Fixing issue Adding support for comparing a string with a datetime
When making a request on a table with a string filter on a datetime column,
the code panics when comparing.

Analysis of the issue:
  when a WHERE datimeColumn < '2020-04-01 11:00:00' is analyzed,
the code does not understand the string as a datetime. It sends this
string to the comparator. The comparator tries to cast the string as
a time.Time and panics.
  I tried to cast the string as a date by using
  WHERE datimeColumn < CAST('2020-04-01 11:00:00' AS DATETIME)
  but this still panics. I tried to cast both of members but still
  the same.

Here are the changes done to fix the problem:
  when the analyzer makes a comparison, there is no ConvertToDatetime
string. There is only one for Timestamp. I have added one but may be
there is a design choice behind this that I'm not aware of. I have also
changed the destination of the ConvertToTimestamp to now be timestamp
and not anymore datetime (since they are different notions).
  the main change is in the Compare function of the datetimeT type.
The idea is the following: the interfaces of the items involved in the comparison
should be transformed to time.Time . In the previous code, there was
no attempt in converting the string to a time.Time. So there code
that now checks the type of the inputs. According to this, a cast is
done or a conversion is attempted. If the comparison succeeds, we
proceed. Else an error is reported.

Unit tests have been added to validate the behavior.

Other potential problems that I have seen:
  * the code stores datetime as UTC which is not what is expected in
    the documentation of MySQL (https://dev.mysql.com/doc/refman/8.0/en/datetime.html)
    This might lead to some consistency issues with MySQL maybe
  * the code stores datetime as UTC but it does not change the time to match
    the original time. For instance:
    you define a time as 2020-04-24 12:00:00 GMT+1. You store in the database and it is
    stored as 2020-04-24T12:00:00Z, which should be instead 2020-04-24T11:00:00Z
    or 2020-04-24T10:00:00Z according to the saving time.
  * From what I have seen in the tests, DATE stores the date and the time. This is not
    what is expected in MySQL so this might lead to inconsistencies between the
    behavior of both.

Equality between string and datetime can be a little bit confusing. The thing
is that a datetime stores the fraction time. So a time as
'2020-04-01 11:00:00' has little or no change to match a real datetime that would
be like 2020-04-01 11:00:00.12345678 . So a user might be confused when searching
for equality but not get it (I was tricked :-) ).
Since the definition of the datetime says that it contains this, I think it is normal
so there should be nothing else to do but may be my changes have impact on choices
that were made at design phasis so I wanted to mention the potential issue.