Fixing issue Adding support for comparing a string with a datetime
When making a request on a table with a string filter on a datetime column,
the code panics when comparing.

Analysis of the issue:
  when a WHERE datimeColumn < '2020-04-01 11:00:00' is analyzed,
the code does not understand the string as a datetime. It sends this
string to the comparator. The comparator tries to convert to something.
And then fails when it compares. The fact that the right element is
analyzed as a string is the first problem.

Trying to cast the string as a date by using
  WHERE datimeColumn < CAST('2020-04-01 11:00:00' AS DATETIME)
still does not work. Even if the comparison this time identifies at first
sight as DATETIME, then it still considers it as string.
And so still fails.

Here are the changes done to fix the problem:
  when the analyzer makes a comparison, first give a chance to the
datetime/string, time/string, timestamp/string possibility. If this
is the case, converting the string to the type of the other element.

To test the different type of TIME types, new methods have been added.
Then, specific conversion strings have been added to represent the way
to convert the types.
May be some were not available by design and then this code might have
miss something.
Code has also changed the destination of the ConvertToTimestamp to now
be dedicated only to timestamp and not anymore datetime (since they
are different notions).

The main change is in the Compare function of the datetimeT type.
The idea is the following: the interfaces of the items involved in the comparison
should be transformed to time.Time . In the previous code, there was
no attempt in converting the string to a time.Time. So there code
that now checks the type of the inputs. According to this, a cast is
done or a conversion is attempted. If the comparison succeeds, we
proceed. Else an error is reported.

Unit tests have been added to validate the behavior.